+++ amethyst_rendy/src/pass/flat2d.rs
@@ -103,7 +103,7 @@ impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2D<B> {
         resources: &Resources,
     ) -> PrepareResult {
         #[cfg(feature = "profiler")]
-        profile_scope!("prepare");
+        profile_scope!("prepare opaque");
 
         let (
             sprite_sheet_storage,
@@ -209,7 +209,7 @@ impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2D<B> {
         _resources: &Resources,
     ) {
         #[cfg(feature = "profiler")]
-        profile_scope!("draw");
+        profile_scope!("draw opaque");
 
         let layout = &self.pipeline_layout;
         encoder.bind_graphics_pipeline(&self.pipeline);
@@ -307,7 +307,7 @@ impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2DTransparent<B> {
         resources: &Resources,
     ) -> PrepareResult {
         #[cfg(feature = "profiler")]
-        profile_scope!("prepare_trans");
+        profile_scope!("prepare transparent");
 
         let (sprite_sheet_storage, tex_storage, visibility, sprite_renders, transforms) =
             <(
@@ -380,7 +380,7 @@ impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2DTransparent<B> {
         _resources: &Resources,
     ) {
         #[cfg(feature = "profiler")]
-        profile_scope!("draw_trans");
+        profile_scope!("draw transparent");
 
         let layout = &self.pipeline_layout;
         encoder.bind_graphics_pipeline(&self.pipeline);