+++ amethyst_test/src/amethyst_application.rs
@@ -130,15 +130,8 @@ where
 {
     /// Returns the built Application.
     ///
-    /// If you are intending to call `.run()` on the `Application` in a test, be aware that on
-    /// Linux, this will cause a segfault when `RenderBundle` is added and GL is using software
-    /// rendering, such as when using Xvfb or when the following environmental variable is set:
-    /// `LIBGL_ALWAYS_SOFTWARE=1`.
-    ///
-    /// To avoid this, please call `.run()` instead of this method, which runs the application in a
-    /// separate thread and waits for it to end before returning.
-    ///
-    /// See <https://users.rust-lang.org/t/trouble-identifying-cause-of-segfault/18096>
+    /// If you are intending to run the `Application`, you can use the `run()` or `run_isolated()`
+    /// methods instead.
     pub fn build(self) -> Result<CoreApplication<'static, GameData<'static, 'static>, E, R>, Error>
     where
         for<'b> R: EventReader<'b, Event = E>,
@@ -886,8 +879,27 @@ mod test {
     /// * That sub-thread is joined.
     /// * Another sub-thread accesses the COM object through the same `lazy_static` variable.
     ///
+    /// To check this:
+    ///
+    /// ```bash
+    /// # On Windows
+    /// cd amethyst_test
+    /// cargo test --features audio
+    /// ```
+    ///
+    /// **Note:** If you simply need an audio file to be loaded, just add a `Processor::<Source>` in
+    /// the test setup:
+    ///
+    /// ```rust,ignore
+    /// use amethyst::{assets::Processor, audio::Source};
+    ///
+    /// AmethystApplication::blank()
+    ///     .with_system(Processor::<Source>::new(), "source_processor", &[])
+    ///     // ...
+    /// ```
+    ///
     /// For more details, see <https://github.com/amethyst/amethyst/issues/1595>.
-    #[cfg(feature = "graphics")]
+    #[cfg(feature = "audio")]
     mod audio_test {
         use amethyst::{
             assets::AssetStorage,
@@ -896,11 +908,10 @@ mod test {
         };
 
         use super::AmethystApplication;
-        use crate::RenderBaseAppExt;
 
         #[test]
         fn audio_zero() -> Result<(), Error> {
-            AmethystApplication::render_base()
+            AmethystApplication::blank()
                 .with_bundle(AudioBundle::default())
                 .with_assertion(|world| {
                     world.read_resource::<AssetStorage<Source>>();
@@ -910,7 +921,7 @@ mod test {
 
         #[test]
         fn audio_one() -> Result<(), Error> {
-            AmethystApplication::render_base()
+            AmethystApplication::blank()
                 .with_bundle(AudioBundle::default())
                 .with_assertion(|world| {
                     world.read_resource::<AssetStorage<Source>>();
@@ -920,7 +931,7 @@ mod test {
 
         #[test]
         fn audio_two() -> Result<(), Error> {
-            AmethystApplication::render_base()
+            AmethystApplication::blank()
                 .with_bundle_fn(|| AudioBundle::default())
                 .with_assertion(|world| {
                     world.read_resource::<AssetStorage<Source>>();
@@ -930,7 +941,7 @@ mod test {
 
         #[test]
         fn audio_three() -> Result<(), Error> {
-            AmethystApplication::render_base()
+            AmethystApplication::blank()
                 .with_bundle_fn(|| AudioBundle::default())
                 .with_assertion(|world| {
                     world.read_resource::<AssetStorage<Source>>();